--- crates/undo/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "undo"
-description = "Subcrate for the editor crate. Contains undo functionality."
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bevy.workspace = true
-pretty-type-name = "1.0.1"
